refactor(api): extract durable run control
